refactor: move KeyFeatures to organisms following atomic design
- Move KeyFeatures from mexico to organisms folder

- Update import path in mexico page

- Maintain data with presentation layer
