+++ app/(site)/mexico.tsx
@@ -5,13 +5,13 @@ import { ContainerLarge } from '~/components/Containers'
 import { HeadInfo } from '~/components/HeadInfo'
 import { FAQSection } from '~/components/mexico/FAQ'
 import { GallerySection } from '~/components/mexico/Gallery'
-import { KeyFeatures } from '~/components/mexico/KeyFeatures'
 import { LocationClimate } from '~/components/mexico/LocationClimate'
 import { MembershipTiers } from '~/components/mexico/MembershipTiers'
 import { NewsUpdates } from '~/components/mexico/NewsUpdates'
 import { ProjectImpact } from '~/components/mexico/ProjectImpact'
 import { ProjectRoadmap } from '~/components/mexico/ProjectRoadmap'
 import { TeamSection } from '~/components/mexico/TeamSection'
+import { KeyFeatures } from '~/components/organisms/KeyFeatures'
 import { HomeH2, HomeH3 } from '~/features/site/home/HomeHeaders'
 import { SocialLinksRow } from '~/features/site/home/SocialLinksRow'
 import { HomeSection, TintSection } from '~/features/site/home/TintSection'
