+++ components/organisms/KeyFeatures.tsx
